legacy/original-files/requirements.txt
requests==2.26.0
reppy==0.4.14
lxml==4.6.2
pyyaml==5.2
tldextract==3.1.2
flask==2.0.2
jieba==0.42.1
tqdm==4.41.1
waitress==2.0.0
guppy3==3.1.2
brotli==1.0.9
urllib3==1.22
rimo-storage==1.3.1
rimo-utils==1.7.0
fasttext==0.9.2
Levenshtein==0.18.1
orjson==3.6.5
persist-queue==0.8.0
numpy==1.22.2
pypinyin==0.47.1
prometheus-client==0.15.0
Werkzeug==2.0.2